requirements.txt
accelerate==1.1.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apturl==0.5.2
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.4
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cos_python_sdk_v5==1.9.41
crcmod==1.7
cryptography==3.4.8
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cupshelpers==1.0
datasets==2.18.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.3.1
fail2ban==0.11.2
fastapi==0.128.0
fasteners==0.14.1
filelock==3.20.3
frozenlist==1.8.0
fsspec==2024.2.0
future==0.18.2
h11==0.16.0
hf-xet==1.2.0
httplib2==0.20.2
huggingface-hub==0.36.0
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==4.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
modelscope==1.18.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
netifaces==0.11.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
olefile==0.46
packaging==26.0
pandas==2.3.3
paramiko==2.9.3
peewee==3.19.0
peft==0.13.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.5.1
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pyarrow==23.0.0
pyarrow-hotfix==0.7
pycairo==1.20.1
pycryptodome==3.23.0
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pyecharts==2.0.9
Pygments==2.19.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvml==13.0.1
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qwen-vl-utils==0.0.8
regex==2026.1.15
reportlab==3.6.8
requests==2.32.5
rich==13.9.4
safetensors==0.7.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentencepiece==0.2.0
simplejson==3.20.2
six==1.16.0
ssh-import-id==5.11
starlette==0.50.0
swanboard==0.1.7b1
swankit==0.1.2b6
swanlab==0.4.5
sympy==1.14.0
systemd-python==234
tokenizers==0.20.3
torch==2.10.0
tqdm==4.67.1
transformers==4.46.2
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.11.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.40.0
wadllib==1.3.6
wcwidth==0.3.5
wrapt==2.0.1
xdg==5
xkit==0.0.0
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
zipp==1.0.0
